// src/app/models/JobInformation.ts
import { SalaryChart } from "./SalaryChart";

export interface JobInformation {
  companyId: number;
  jobId: number;
  jobTitle: string;
  jobPostingDate: Date;
  proceedToPublishDate: Date;
  deadlineDate: Date;
  isExpired: boolean;
  publishDate: Date;
  isPaused: boolean;
  jobPaymentStatus: boolean;
  serviceType: string;
  isPublished: boolean;
  jobViewedCount: number;
  jobViewedCountForJobSummary: number;
  jobViewedCountForJobDetails: number;
  salesVerify: number;
  accountsVerify: number;
  applyOptions: string;
  contentVerify: number;
  totalApplications: number;
  applicationViewedCount: number;
  applicationYetToViewCount: number;
  applicationShortlistedCount: number;
  lastActivityName: string;
  profileMatchingScoreBelow50: number;
  profileMatchingScore50To69: number;
  profileMatchingScore70To89: number;
  profileMatchingScoreAtLeast90: number;
  experience0To2Years: number;
  experience2To3Years: number;
  experience3To5Years: number;
  experience5To10Years: number;
  experienceMoreThan10Years: number;
  employeerGivenSalaryRange: string;
  belowEmployeerGivenSalaryRange: number;
  applicantCountsWithinEmployeerGivenSalaryRange : number;
  greaterThanEmployeerGivenSalaryRange: number;
  readMessageCount: number;
  unreadMessageCount: number;
  isDrafted: boolean;
  hasContactId: boolean;
  notificationCount: number;
  notificationDate: Date;
  messageCount: number;
  messageDate: Date;
  isJobVerified: boolean;
  lastActivityApplicantsCount: number;
  salaryChartData: SalaryChart[];
  isForwardToNewLink: number;
  regionalJob: number;
}